db+sqlite+pg: split address types into per-op files
Move ListAddressTypes/GetAddressType into addressstore_types.go in each
package. These are thin reads (one query plus row conversion), so per the
wallet/internal/db README they use direct backend implementations rather
than an ops interface: each backend runs its own sqlc query, row conversion,
and not-found mapping.
